Module type Conback.CONSOLE

include Mirage_console.S

include Mirage_device.S

type t

The type representing the internal state of the device

val disconnect : t -> unit Lwt.t

Disconnect from the device. While this might take some time to complete, it can never result in an error.

include Mirage_flow.S with type flow := t

type error

The type for flow errors.

val pp_error : error Fmt.t

pp_error is the pretty-printer for errors.

type nonrec write_error#row

The type for write errors.

and write_error = private [>

| Mirage_flow.write_error

]

The type for write errors.

val pp_write_error : write_error Fmt.t

pp_write_error is the pretty-printer for write errors.

val read : t -> (Cstruct.t Mirage_flow.or_eof, error) Stdlib.result Lwt.t

read flow blocks until some data is available and returns a fresh buffer containing it.

The returned buffer will be of a size convenient to the flow implementation, but will always have at least 1 byte.

If the remote endpoint calls close then calls to read will keep returning data until all the in-flight data has been read. read flow will return `Eof when the remote endpoint has called close and when there is no more in-flight data.

val write : t -> Cstruct.t -> (unit, write_error) Stdlib.result Lwt.t

write flow buffer writes a buffer to the flow. There is no indication when the buffer has actually been read and, therefore, it must not be reused. The contents may be transmitted in separate packets, depending on the underlying transport. The result Ok () indicates success, Error `Closed indicates that the connection is now closed and therefore the data could not be written. Other errors are possible.

val writev : t -> Cstruct.t list -> (unit, write_error) Stdlib.result Lwt.t

writev flow buffers writes a sequence of buffers to the flow. There is no indication when the buffers have actually been read and, therefore, they must not be reused. The result Ok () indicates success, Error `Closed indicates that the connection is now closed and therefore the data could not be written. Other errors are possible.

val close : t -> unit Lwt.t

close flow flushes all pending writes and signals the remote endpoint that there will be no future writes. Once the remote endpoint has read all pending data, it is expected that calls to read on the remote return `Eof.

Note it is still possible for the remote endpoint to write to the flow and for the local endpoint to call read. This state where the local endpoint has called close but the remote endpoint has not called close is similar to that of a half-closed TCP connection or a Unix socket after shutdown(SHUTDOWN_WRITE).

close flow waits until the remote endpoint has also called close before returning. At this point no data can flow in either direction and resources associated with the flow can be freed.

val log : t -> string -> unit Lwt.t

log t str writes str to the console t, appending appropriate line endings. If close was called on the console before, str is discarded silently.

val connect : string -> t Lwt.t